pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flows", "utils"]

[project]
name = "orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "clickhouse-connect>=0.8.18",
    "jinja2>=3.1.6",
    "pandas>=2.3.0",
    "polars>=1.32.3",
    "prefect>=3.4.6",
    "prefect-aws>=0.5.10",
    "prefect-docker>=0.6.6",
    "prefect-shell>=0.3.1",
    "python-dotenv>=1.1.0",
    "soundcharts>=0.0.6",
]
